#89: updated Experiments page to three sections
