+++ src/services/fakeSlices.js
@@ -1,0 +1,29 @@
+const slices = [
+  {
+    "graph_id": "1",
+    "lease_end": "2021-09-05 11:30:25",
+    "slice_id": "1",
+    "slice_name": "1-site-slice",
+    "slice_state": "StableOK"
+  },
+  {
+    "graph_id": "2",
+    "lease_end": "2021-09-02 14:56:57",
+    "slice_id": "2",
+    "slice_name": "2-site-slice",
+    "slice_state": "StableOK"
+  }
+];
+
+const sliceProperties = {
+  "1": {"directed": false, "multigraph": false, "graph": {}, "nodes": [{"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkNode", "NodeID": "70406bb2-44be-4d72-a315-03355d21622d", "Name": "n1", "Type": "VM", "Site": "RENC", "StitchNode": "false", "id": 1}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkNode", "NodeID": "e8b81c32-4ffd-4cbf-aa0a-7387cee1fc95", "Name": "n2", "Type": "VM", "Site": "RENC", "StitchNode": "false", "id": 2}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkNode", "NodeID": "3a77bb46-f78a-454d-82b0-877855cec238", "Name": "n3", "Type": "VM", "Site": "RENC", "StitchNode": "false", "id": 3}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Component", "NodeID": "2e026901-78dc-4b08-9241-a1b8b88be36c", "Name": "nic1", "Type": "SharedNIC", "Model": "ConnectX-6", "Details": "Mellanox ConnectX-6 VPI MCX653 dual port 100Gbps", "StitchNode": "false", "id": 4}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkService", "NodeID": "1d2e60bd-8fec-47ec-a9b0-c673131b3192", "Name": "n1-nic1-l2ovs", "Type": "OVS", "StitchNode": "false", "Layer": "L2", "id": 5}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "4ef4ad5f-4149-447d-98aa-09af10e88a05", "Name": "nic1-p1", "Type": "SharedPort", "Capacities": "{\"bw\": 100, \"unit\": 1}", "StitchNode": "false", "id": 6}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Component", "NodeID": "04bbdae6-71ce-46bd-8604-04bb2622a30d", "Name": "drive1", "Type": "NVME", "Model": "P4510", "Details": "Dell Express Flash NVMe P4510 1TB SFF", "StitchNode": "false", "id": 7}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Component", "NodeID": "164cd91c-4674-4ac8-89d4-c349df049b6f", "Name": "nic1", "Type": "SmartNIC", "Model": "ConnectX-6", "Details": "Mellanox ConnectX-6 VPI MCX653 dual port 100Gbps", "StitchNode": "false", "id": 8}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkService", "NodeID": "7808d6b2-57a0-4adf-bfe9-e90005327ea2", "Name": "n2-nic1-l2ovs", "Type": "OVS", "StitchNode": "false", "Layer": "L2", "id": 9}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "69802842-72f9-4091-b41f-3b7a4c814adc", "Name": "nic1-p1", "Type": "DedicatedPort", "Capacities": "{\"bw\": 100, \"unit\": 1}", "StitchNode": "false", "id": 10}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "4b5b53a3-9e25-4af7-adcb-c963ca329631", "Name": "nic1-p2", "Type": "DedicatedPort", "Capacities": "{\"bw\": 100, \"unit\": 1}", "StitchNode": "false", "id": 11}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Component", "NodeID": "720a6a28-e85b-40db-9c73-1cbe7cc34fc9", "Name": "gpu1", "Type": "GPU", "Model": "RTX6000", "Details": "NVIDIA Corporation TU102GL [Quadro RTX 6000/8000] (rev a1)", "StitchNode": "false", "id": 12}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Component", "NodeID": "5efcc24b-7fab-47b4-b8d2-8c14ef1f47a2", "Name": "nic1", "Type": "SmartNIC", "Model": "ConnectX-5", "Details": "Mellanox ConnectX-5 Dual Port 10/25GbE", "StitchNode": "false", "id": 13}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkService", "NodeID": "64e56a4d-6aa8-4bd3-8125-9d88f5594939", "Name": "n3-nic1-l2ovs", "Type": "OVS", "StitchNode": "false", "Layer": "L2", "id": 14}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "5097e6bc-7d40-4450-881f-2ea570688d7a", "Name": "nic1-p1", "Type": "DedicatedPort", "Capacities": "{\"bw\": 25, \"unit\": 1}", "StitchNode": "false", "id": 15}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "7ed13fb3-3734-4c52-9a79-a5560ddfcd45", "Name": "nic1-p2", "Type": "DedicatedPort", "Capacities": "{\"bw\": 25, \"unit\": 1}", "StitchNode": "false", "id": 16}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "NetworkService", "NodeID": "622435d0-16ed-476a-bdc0-4b406cbcc9a5", "Name": "bridge1", "Type": "L2Bridge", "Site": "RENC", "StitchNode": "false", "Layer": "L2", "id": 17}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "1f260cf8-272e-4458-b080-b8a30fb8a581", "Name": "bridge1-nic1-p1", "Type": "ServicePort", "StitchNode": "false", "id": 18}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Link", "NodeID": "a8bd486f-7249-4af0-bac1-e88aa4f2e7be", "Name": "bridge1-nic1-p1-link", "Type": "L2Path", "StitchNode": "false", "Layer": "L2", "id": 19}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "232ba66f-2e6a-46e8-8002-77f8c5d85961", "Name": "bridge1-nic1-p1", "Type": "ServicePort", "StitchNode": "false", "id": 20}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Link", "NodeID": "0237a120-e4fc-40db-92c0-43211f8c0c2a", "Name": "bridge1-nic1-p1-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 21}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "88377e6a-e253-4228-87b9-2fee2f3ea3d6", "Name": "bridge1-nic1-p2", "Type": "ServicePort", "StitchNode": "false", "id": 22}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Link", "NodeID": "264fb240-1102-434f-81f9-4e15f57f146c", "Name": "bridge1-nic1-p2-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 23}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "00f620a2-8bd6-45f6-a086-ff8e9fcc5a39", "Name": "bridge1-nic1-p2", "Type": "ServicePort", "StitchNode": "false", "id": 24}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Link", "NodeID": "ba0e477f-74ae-466c-8ec3-b60a3134e781", "Name": "bridge1-nic1-p2-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 25}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "ConnectionPoint", "NodeID": "0ff14e2a-5813-49e7-82f9-ece677ff0510", "Name": "bridge1-nic1-p1", "Type": "ServicePort", "StitchNode": "false", "id": 26}, {"GraphID": "83ac87e1-a6ef-45e2-a01b-00ba3d31daf7", "Class": "Link", "NodeID": "6275a6f8-ae48-4fa7-b999-7e87c286271d", "Name": "bridge1-nic1-p1-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 27}], "links": [{"Class": "has", "source": 1, "target": 4}, {"Class": "has", "source": 1, "target": 7}, {"Class": "has", "source": 2, "target": 8}, {"Class": "has", "source": 2, "target": 12}, {"Class": "has", "source": 3, "target": 13}, {"Class": "has", "source": 4, "target": 5}, {"Class": "connects", "source": 5, "target": 6}, {"Class": "connects", "source": 6, "target": 19}, {"Class": "has", "source": 8, "target": 9}, {"Class": "connects", "source": 9, "target": 10}, {"Class": "connects", "source": 9, "target": 11}, {"Class": "connects", "source": 10, "target": 21}, {"Class": "connects", "source": 11, "target": 23}, {"Class": "has", "source": 13, "target": 14}, {"Class": "connects", "source": 14, "target": 15}, {"Class": "connects", "source": 14, "target": 16}, {"Class": "connects", "source": 15, "target": 27}, {"Class": "connects", "source": 16, "target": 25}, {"Class": "connects", "source": 17, "target": 18}, {"Class": "connects", "source": 17, "target": 20}, {"Class": "connects", "source": 17, "target": 22}, {"Class": "connects", "source": 17, "target": 24}, {"Class": "connects", "source": 17, "target": 26}, {"Class": "connects", "source": 18, "target": 19}, {"Class": "connects", "source": 20, "target": 21}, {"Class": "connects", "source": 22, "target": 23}, {"Class": "connects", "source": 24, "target": 25}, {"Class": "connects", "source": 26, "target": 27}]},
+  "2": {"directed": false, "multigraph": false, "graph": {}, "nodes": [{"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkNode", "NodeID": "11d48e7b-ed0b-4d57-b81e-397f2354e9fa", "Name": "n1", "Type": "VM", "Site": "RENC", "StitchNode": "false", "id": 28}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkNode", "NodeID": "488599ad-3a4f-4165-b5d8-6bbc16f6c657", "Name": "n2", "Type": "VM", "Site": "RENC", "StitchNode": "false", "id": 29}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkNode", "NodeID": "455d6dcf-eef4-4f49-9901-3ec60c657fda", "Name": "n3", "Type": "VM", "Site": "UKY", "StitchNode": "false", "id": 30}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Component", "NodeID": "a89e70b1-8d8d-48bf-aaf7-ff8b26d06165", "Name": "nic1", "Type": "SharedNIC", "Model": "ConnectX-6", "Details": "Mellanox ConnectX-6 VPI MCX653 dual port 100Gbps", "StitchNode": "false", "id": 31}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkService", "NodeID": "95b6f873-2d67-40e9-988d-ae694c2e83ee", "Name": "n1-nic1-l2ovs", "Type": "OVS", "StitchNode": "false", "Layer": "L2", "id": 32}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "38ad48db-d970-4a96-a6e7-2386b0d597ba", "Name": "nic1-p1", "Type": "SharedPort", "Capacities": "{\"bw\": 100, \"unit\": 1}", "StitchNode": "false", "id": 33}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Component", "NodeID": "ac503180-229b-4d2e-92e2-d36129884e02", "Name": "drive1", "Type": "NVME", "Model": "P4510", "Details": "Dell Express Flash NVMe P4510 1TB SFF", "StitchNode": "false", "id": 34}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Component", "NodeID": "d6acabb2-1e5a-445f-8155-96639d6252cd", "Name": "nic1", "Type": "SmartNIC", "Model": "ConnectX-6", "Details": "Mellanox ConnectX-6 VPI MCX653 dual port 100Gbps", "StitchNode": "false", "id": 35}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkService", "NodeID": "ec8db655-a21b-46b9-828c-d86517b115cd", "Name": "n2-nic1-l2ovs", "Type": "OVS", "StitchNode": "false", "Layer": "L2", "id": 36}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "6d7b64bb-d0c9-468a-851b-4451384d3777", "Name": "nic1-p1", "Type": "DedicatedPort", "Capacities": "{\"bw\": 100, \"unit\": 1}", "StitchNode": "false", "id": 37}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "8d5d3c77-550c-4cb9-bd82-f5e8e83eb403", "Name": "nic1-p2", "Type": "DedicatedPort", "Capacities": "{\"bw\": 100, \"unit\": 1}", "StitchNode": "false", "id": 38}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Component", "NodeID": "89d497a7-339f-4817-b62b-f33f31f3a71d", "Name": "gpu1", "Type": "GPU", "Model": "RTX6000", "Details": "NVIDIA Corporation TU102GL [Quadro RTX 6000/8000] (rev a1)", "StitchNode": "false", "id": 39}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Component", "NodeID": "d75b11ab-03d8-4c5c-b8c8-7cc92055106c", "Name": "nic1", "Type": "SmartNIC", "Model": "ConnectX-5", "Details": "Mellanox ConnectX-5 Dual Port 10/25GbE", "StitchNode": "false", "id": 40}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkService", "NodeID": "ad63c7af-3e43-4c23-b02c-6a2953d3a17a", "Name": "n3-nic1-l2ovs", "Type": "OVS", "StitchNode": "false", "Layer": "L2", "id": 41}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "8662ccf9-c6f3-430b-9e76-8bd3bb3071f2", "Name": "nic1-p1", "Type": "DedicatedPort", "Capacities": "{\"bw\": 25, \"unit\": 1}", "StitchNode": "false", "id": 42}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "283eb2a8-fd33-433d-a7a8-0a27275bd62e", "Name": "nic1-p2", "Type": "DedicatedPort", "Capacities": "{\"bw\": 25, \"unit\": 1}", "StitchNode": "false", "id": 43}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "NetworkService", "NodeID": "751d57cd-b60d-4b1a-9380-cc8b233f7aef", "Name": "bridge1", "Type": "L2STS", "StitchNode": "false", "Layer": "L2", "id": 44}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "0372e5ba-2fcf-4d40-b9b0-2f08c994ea66", "Name": "bridge1-nic1-p1", "Type": "ServicePort", "StitchNode": "false", "id": 45}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Link", "NodeID": "6f75a0d9-82d2-4262-b968-222bcd8e82e1", "Name": "bridge1-nic1-p1-link", "Type": "L2Path", "StitchNode": "false", "Layer": "L2", "id": 46}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "4af351a5-521d-4396-97dd-a30a88bd28e8", "Name": "bridge1-nic1-p1", "Type": "ServicePort", "StitchNode": "false", "id": 47}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Link", "NodeID": "ffc3558e-701e-4fe8-9aeb-00453a8691cc", "Name": "bridge1-nic1-p1-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 48}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "d98e70a4-c409-4c7b-8d4b-66cfe4ee2c33", "Name": "bridge1-nic1-p2", "Type": "ServicePort", "StitchNode": "false", "id": 49}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Link", "NodeID": "d93a63fc-4ba1-4b47-a6ae-3a092334a9bc", "Name": "bridge1-nic1-p2-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 50}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "f91b019a-3dd7-44f3-92b4-7b941c9c3c5c", "Name": "bridge1-nic1-p1", "Type": "ServicePort", "StitchNode": "false", "id": 51}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Link", "NodeID": "b18f6b29-70af-4952-a2d6-26f675579ae0", "Name": "bridge1-nic1-p1-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 52}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "ConnectionPoint", "NodeID": "bdc6450b-11f0-41bd-be6e-586046435ead", "Name": "bridge1-nic1-p2", "Type": "ServicePort", "StitchNode": "false", "id": 53}, {"GraphID": "5424bb7d-edbf-4d12-8472-7a4fa58f5ecd", "Class": "Link", "NodeID": "ff26fcfc-4a7f-4482-9bfd-e28f311d2d7f", "Name": "bridge1-nic1-p2-link", "Type": "Patch", "StitchNode": "false", "Layer": "L2", "id": 54}], "links": [{"Class": "has", "source": 28, "target": 31}, {"Class": "has", "source": 28, "target": 34}, {"Class": "has", "source": 29, "target": 35}, {"Class": "has", "source": 29, "target": 39}, {"Class": "has", "source": 30, "target": 40}, {"Class": "has", "source": 31, "target": 32}, {"Class": "connects", "source": 32, "target": 33}, {"Class": "connects", "source": 33, "target": 46}, {"Class": "has", "source": 35, "target": 36}, {"Class": "connects", "source": 36, "target": 37}, {"Class": "connects", "source": 36, "target": 38}, {"Class": "connects", "source": 37, "target": 48}, {"Class": "connects", "source": 38, "target": 50}, {"Class": "has", "source": 40, "target": 41}, {"Class": "connects", "source": 41, "target": 42}, {"Class": "connects", "source": 41, "target": 43}, {"Class": "connects", "source": 42, "target": 52}, {"Class": "connects", "source": 43, "target": 54}, {"Class": "connects", "source": 44, "target": 45}, {"Class": "connects", "source": 44, "target": 47}, {"Class": "connects", "source": 44, "target": 49}, {"Class": "connects", "source": 44, "target": 51}, {"Class": "connects", "source": 44, "target": 53}, {"Class": "connects", "source": 45, "target": 46}, {"Class": "connects", "source": 47, "target": 48}, {"Class": "connects", "source": 49, "target": 50}, {"Class": "connects", "source": 51, "target": 52}, {"Class": "connects", "source": 53, "target": 54}]},
+}
+
+export function getSlices() {
+  return slices;
+}
+
+export function getSliceById(uuid) {
+  return sliceProperties[uuid];
+}
